+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog 
 
 ## HEAD
-
+- fix properly export genesis state
+- fix properly init genesis state from old state
+- fix iterate all domains
+- add iterate all accounts
+- change one store key for domain
+- change shorter indexing keys
+- add MsgSetAccountURI
 ## 0.2.0
 - change path prefix to star